Karsha
Karsha is a new way of online/offline payments.A user who wants to use
karsha first registers as a customer in the site.He then logs in to
select the desired product and then goes on to generate a unique Karsha
Id(also known as the K-id).This K-id is a uniquely generated 16 bit
number.This Id holds all the details of the current transaction.At the
same time,this id is also generated in the merchant cloud.

When the user proceeds,the K-ids of both user and customer clouds is
matched with and the result of the transaction is known to the
user.Further the user goes on to generate the K-note which contains the
K-id and a unique bar code.This K-note can be used for making offline
payments replacing traditional cash/credit cards.Digitialized version of
K-note can also be used for offline payments using bluetooth/NFC.Thus
the user is able to make online/offline payments with minimal
authentication and in a more secured way.